+++ src/redturtle/prenotazioni/config.py
@@ -5,7 +5,6 @@ PROJECTNAME = "redturtle.prenotazioni"
 MIN_IN_DAY = 24 * 60
 PAUSE_SLOT = "pause_slot"
 PAUSE_PORTAL_TYPE = "DayPause"
-DELETE_TOKEN_KEY = "delete_token"
 VERIFIED_BOOKING = "redturtle.prenotazioni.verified_booking"
 NOTIFICATIONS_LOGS = "redturtle.prenotazioni.notifications_logs"
 REQUIRABLE_AND_VISIBLE_FIELDS = [